add mechanism to load syntax highlighting
Summary:
Use textmate grammars and textmate-lib to load syntax highlighting for a given patch in the comparison view.

This diff merely loads the tokenized ranges, it does not yet render them

We should also actually be doing this syntax highlighting in a webworker or something so we don't block the UI on syntax highlighting. I'll do this in a later diff, just setting things up now.

I also have to import the themes from vscode, because even in vscode we only have UI theme colors and not syntax ones.

Reviewed By: muirdm

Differential Revision: D48377893

fbshipit-source-id: 3bec87d4cf2dde97564f5125b919016d25453686