Render syntax highlighting
Summary:
Now that we've loaded the tokenized lines, we can actually render those ranges. To do this we use the same code as reviewstack, which I moved to shared/.

This is still slightly messy the way it is in this diff, since we have to keep track of indices correctly to make the patch offsets work. Reviewstack uses line numbers to make their patches absolutely positioned, but that's because they fetch the entire file  contents before/after to do their tokenization—we only really have the patch with some limited context, and we load additional lines separately.

Some things remaining to do:
- syntax highlighting for context lines that are loaded later (it's not part of tokenization of the patch... so we need to re-run the tokenization on that loaded part...)
- again, we should probably move this to a web worker... though I will say it's pretty fast in my testing.
- look into the performance a bit, and make sure we're not overrendering or double-highlighting or something (kind of easy to do with effects)
- make this work in vscode and not just browser

Reviewed By: muirdm

Differential Revision: D48377892

fbshipit-source-id: 2c4ddff3a8c5f3fb8077b41c22368cdf23cd4c45