Make syntax highlighting work in vscode extension
Summary:
For syntax highlighting to work, we need to be able to load the textmate grammars and the onigurama webassembly code in the vscode webview.

In browser ISL, this is done by serving these files from `isl/public/`.

The vscode webview build process does not use `isl/public`. So we need to:
1. copy `public/generated/textmate` to the webpack output folder's `webview/genereated/textmate` folder. We do this with a new webpack plugin to copy files (During `yarn watch-webview` or `yarn build-webview`).
2. ensure fetches use proper URI to allow fetching from the webview. The change is simple: we set the `<base>` url in Tthe webview HTML, so just make the urls we fetch relative to the base by not starting with `/`.
3. Adjust the content security policy of the webview HTML to allow the files to be fetched and the webassembly to be loaded. Unfortunately, webassembly requires using `wasm-unsafe-eval` as a CSP. This is technically a bad practice, but it's seemingly the only way to actually load wasm while having a CSP.

Yes, copying these files means all 4MB of grammars and onig.wasm will be part of our extension bundle size. I think it's worth it for syntax highlighting.

Reviewed By: muirdm

Differential Revision: D48411170

fbshipit-source-id: 840771d9808dfb9be1f9d9900d63ffaad4da121c