Fix bug in `DeletionLog::update_candidates`
Summary:
With this bug, if updating the deletion log in multiple chunks, we would only update the first chunk.

That is because we were comparing the number of blobs already in the target state in the entire DB with the number of blobs in this chunk only.

For instance, if iterating over chunks of size 1, we would call `update_candidates` once with `cs_to_blobs` of size 1.
At this stage, the number of blobs `in_target_state` in the db would be 0 and the number of blobs `not_in_target_state` would be N (total number of blobs in the DB for this reason).
The `blobs_count` would be 1.

So we would update the first chunk.

When hitting the next chunk, the number of blobs `in_target_state` would be 1 which would be equal to the `blobs_count`, so we would early exit.

The intent of that code was to check whether the blobs **that we are trying to update at the moment** are in the target state, so we need the extra filtering that is added in this diff.

Reviewed By: markbt

Differential Revision: D48429996

fbshipit-source-id: 8208de2b217044f5b590682e23676affab875375