status: write file mode and size to treestate
Summary:
Now the Rust status also writes files' mode and size back to the treestate (in addition to mtime). This is an optimization that avoids future status work for clean files. In particular, if a distate entry has no size available and is marked NEED_CHECK, we must go to disk to compare file contents. If the file is clean and we don't write the size back to the dirstate, we could have to go back to disk again in the future.

The only way I know that a dirstate entry can "lose" its size and mode is via the "debugrebuilddirstate" command, so in practice I don't think it is very important, but there may be other ways I don't know about.

Reviewed By: quark-zju

Differential Revision: D48342313

fbshipit-source-id: 7f0a2fdbc81bafd27e4624d926f01aba446c879a