Mark changesets created by lossy conversion as such
Summary:
Put this feature behind a tunable for ease of rollout and revert.
For safety, we can roll it out while the repos are locked.

During cross repo sync, if a changeset is created where not all file changes from the original changeset have an equivalent file change in the generated changeset, the conversion is said to be lossy.

In that case, set a new flag: `created_by_lossy_conversion` in the `hg_extra` metadata to indicate this fact to `hg` and `sapling`.

This can be used to prevent users from performing unsound operations such as `backout` on such a lossy changeset. Instead, they should perform the `backout` in the source-of-truth repo and let it be backsynced into the destination repo.

For more context, see [this post](https://fburl.com/workplace/w7f8gtsh)

Reviewed By: YousefSalama

Differential Revision: D48470112

fbshipit-source-id: 2a557239189c2fb997665d9f464bebe6c4b159f3