tests: support multiple features in expected lines
Summary:
As the title mentions. This should make the line to be expected only if all the features for the line are enabled.

For instance, if one has
```
dir/l -> yadda (symlink !) (unix-permissions !)
```

that line will only be expected if the test has both the `symlink` and `unix-permissions` features enabled

Reviewed By: muirdm

Differential Revision: D48489317

fbshipit-source-id: 7ada0743d0e712b8c9c16d5d178230b8eb0e32d9