symlinks: make checkpathconflicts wonk on Windows
Summary:
Previously the only test that tested the `experimental.merge.checkpathconflicts=True` config required symlinks. As we are enabling more symlink tests on D48435504 for Windows, this test needs to be fixed before doing so.

Besides making this config works on Windows, this changes `test-pathconflicts-merge.t` so that the bits of it that require symlinks are surrounded of `#require symlinks`

Reviewed By: muirdm

Differential Revision: D48435503

fbshipit-source-id: dab21b46597a7322620e5ca0a5316c24679a6e3f