symlinks: make file symlinks appear on checkout and clone on Windows
Summary:
Allows the `goto` and `clone` commands (among others) to create symlinks on Windows.

This also makes most of `test-symlinks.t` pass on Windows , but only with the old `status` command. This test was changed temporarily to the old one and will be reverted to the new one when symlink support for the new `status` command is added on Windows.

Reviewed By: muirdm

Differential Revision: D47654981

fbshipit-source-id: e828da6a384f4702607f7ee784ef2285fb044bad