Properly mock execa in tests
Summary:
In the next diff, we use execa for more than just stdout, so we need to improve how our mocking works in our tests so that it can handle calling `.kill` and `.on('exit')`.

Might as well clean things up at the same time, so I added a nicer way to specify commands we intend to mock, with their output or error.

Reviewed By: quark-zju

Differential Revision: D48571932

fbshipit-source-id: 433125aad7a55b16f8840f491bbb55b23b301860