pathmatcher: fix handling of empty "listfile" pattern
Summary:
An empty listfile expands to and empty list of patterns, which was being turned into an always matcher.

To fix, I added an Option<> around the patterns to explicitly differentiate "the user specified zero patterns" and "the user specified patterns but they expanded to nothing".

I re-organized HintedMatcher::new to handle all the "always matches" and "never matches" logic at the start. This simplifies things overall, I think.

Reviewed By: zzl0

Differential Revision: D48459620

fbshipit-source-id: 9306bceb9e07ef3d213a9df5d3e4ada5f71685cc