add amend --to in context menu
Summary:
Add a quick way to amend your selected changes down to a specific commit in the stack. You just right click a commit in your stack, and click "amend changes to here".

This is only allowed in certain situations:
- the commit you click on is a parent of the head commit
- you have uncommitted changes to amend
- you don't have a partial chunk selection (entire file selected)

This involved adding a way to get a read-only view of the partial selection state, since we need to determine the eligibility of the amend --to from a recoil snapshot we have in our callback. I use some typescript magic to get at the `uncommittedSelection` state without requiring the use of useUncommittedSelection, while still being typesafe so you don't try to use the methods which mutate the selection (They would have no effect when used directly from uncommittedSelection). This unfortunately requires listing out all the readonly function names. I couldn't find a way to annotate them more directly.

Reviewed By: quark-zju

Differential Revision: D48590064

fbshipit-source-id: 705f7442f0a9c0e6f9aca6b27ff7a408c8202c78