webview-app: spawn a "webview" application
Summary:
This crate handles the cross-platform complexities of opening a single window
browser of a given URL.

On macOS, create an app bundle (so we have an icon), and launch the app which
opens a window using webview-sys.

On other platforms, try to locate a chrome-like browser, and launch it in the
app mode.

Modify hgmain to make the app bundle entry point work.

Reviewed By: evangrayk

Differential Revision: D48592578

fbshipit-source-id: 831e11d7a8bb466f5285595f6d5e56a8f9611240