status: fix ignored file handling for sparse
Summary: Fix Rust status to treat files outside the sparse profile as ignored. Previously it was treating them as regular files that never matched the matcher. The only distinction I'm aware of is that the latter would always "track" them in the treestate as untracked files, where the former treats them as ignored files (i.e. only tracks if fsmonitor.track-ignore-files=true).

Reviewed By: zzl0

Differential Revision: D48380152

fbshipit-source-id: a4a8ab19f91cdc0d29f282a577e61a42e41bdab8