Use rtl for file paths to have ellipsis on left
Summary:
Un-revert a previous change so we use `text-direction: rtl` again in CSS for file paths. Combined with `text-overflow: ellipsis`, this means we truncate long paths on the left, so you always see the path basename / file extension.

We previously reverted this because it behaved weirdly for paths like `.gitignore`, which would render as `gitignore.` due to rtl.

However, we can work around this with unicode left-to-right markers `\u200E`. It seems we only need a single marker at the start of the string to fix the "gitignore." issue, while still overflowing to the left.

This introduces two problems:
- copying the file path via selection now has \u200E. So we should strip these characters when copying file paths.
- all our tests for file paths fail because they don't expect LTR markers. So we fix all these tests.

Reviewed By: quark-zju

Differential Revision: D48615379

fbshipit-source-id: b77b9814c815ace8ffb524c4920c68b7d4908770