stack: add a more general version of FileStackState.moveLines
Summary:
The `moveLines` API is inflexible: it can move only one chunk in one rev.
We'd like a way to move multiple chunks (if the user selects them) in multiple
revs (if the user selects both `-` and `+` lines in a unified diff view, then
the `-` and `+` lines are from different revs).

This diff adds a more general purposed `mapAllLines` to do that.

Reviewed By: evangrayk

Differential Revision: D48491777

fbshipit-source-id: 9c1aa2dfc386236e4461ea86f583bd0793b16051