Rewrite stripped commits to temp repo
Summary:
## This Stack
See D48563980.

## This diff
Implement the code that actually writes the relevant commits to the temp repo containing only changes to the exported paths.

The functions being implemented expect a list of topologically sorted changesets and a hashmap mapping changeset ids to their parents, which I'm referring to as "partial graph information".

As I mentioned on D48563980, currently the data being create works only for a single path and expects a linear history with no branches (e.g. the example in D48563980).

## Next steps
Not strictly in that order:
- Start looking on how I can export the temporary repo to a git repo. IIUC RajivTS is a good POC for this?
- Look into `list_file_history` to see how I can properly get a proper "history graph", i.e. the commit graph containing only the relevant commits for the exported paths. IIUC YousefSalama is the best POC to help me with this.
- Update the integration tests to use the tool and test what I already implemented.

Differential Revision: D48558621

fbshipit-source-id: 8aba4b8927e9c277b0338f8ae3e66aaad1a2c355