Tests validating DeltaInstructions header encoding
Summary: The tests introduced so far validate that the `DeltaInstructions` get encoded properly and that we are able to regenerate the new object from the base object using these instructions. The two tests added as part of this diff validate that the headers get encoded in the right order and using the correct encoding scheme. It relies on helper functions from the `gix-pack` crate that haven't been exposed yet.

Differential Revision: D48600476

fbshipit-source-id: 37725e251246eba4dcf9d492845cf279127c8e3b