Generating DeltaInstructions by diffing Git objects
Summary: Now that we can create individual `DeltaInstruction`, this diff implements the generation of entire sequence of `DeltaInstructions` to create the `new_object` out of the `base_object`. It uses the `gix-diff` library to provide the diffing functionality. The next diff in the stack will connect the dots from Git objects to Deltas and add unit tests to validate the created instructions work as expected. E2E integration tests will come further down the line.

Differential Revision: D48517226

fbshipit-source-id: 8c80d3a60d754cd2bcbc2442c07559bdeafb8c96