Introduce DeltaInstruction representing Git object Delta
Summary: Since we plan to use `DeltaObjects` instead of full Git objects in packing and bundling, we need a way to represent the steps that need to be followed to regenerate the new object out of the delta of the base object. These steps also need to be Zlib encoded in an efficient format to make the use of deltas viable. This diff introduces `DeltaInstructions` which achieve this objective. Currently they are being tested via very basic unit tests, but the whole E2E validation would happen in a couple of diffs.

Differential Revision: D48494621

fbshipit-source-id: 7791972dc189026d6422d83f9cde813533afd721