Replace println! with slogs
Summary:
Just replacing the `println!` statements with slogs. Almost all of them are `debug` logs, many of which I'll remove later but are still helpful for debugging.

Besides replacing the print statements, I'm also making a small change to basically fail silently when creating the master bookmark.

Instead of crashing execution, this will just log an error for now.

I intend to remove this before production, but it's helpful for now because I keep using an existing repo on my local server for testing and it will fail if I forget to delete the master bookmark.

I thought about just calling the `move_bookmark` function in this case, but I don't think it makes sense to do that in prod because we'll be creating a new repo from scratch, so we shouldn't expect any existing bookmarks...

## What's next
In the new few diffs I'll start using the binary in the integration test and will also add unit tests for the code added in D48558621.

After that, I'll work on exporting the temp repo to a git repo.

Reviewed By: markbt

Differential Revision: D48594816

fbshipit-source-id: d1e107ba6a4475444de93868510dee2da5f901d8