status: enable Rust status for test-lfs.t
Summary:
This test is for the Python lfs extension. Most of the test is happy with the revlog<->eagerstore compat layer (maybe by accident), but one interesting test with renames wasn't working. Fix by using a remotefilelog repo with remotefilelog.lfs enabled, which can coexist with the LFS extension.

Hopefully we can remove the legacy LFS extension soon.

Reviewed By: quark-zju

Differential Revision: D48342319

fbshipit-source-id: d25148f3fcd8f6d99fcf4eba0bbc0d1c09d17e69