merge: fix merge of exec bit change
Summary:
After switching to the Rust status, test-issue1802.t failed. It is testing that the exec bit does not get lost when performing a merge on a platform that doesn't support the exec bit.

The original fix for this case was to infer the exec bit from the parent manifest(s) instead of looking at the disk. However, the Rust status mostly doesn't look at the p2 manifest at all, instead depending on the EXIST_P2 treestate flag to infer a file should be marked modified when merging. So to fix for the Rust status, I tweaked (fixed?) merge.py to mark a flag-only-merge as EXIST_P2. Then Rust naturally picks it up as a modified file.

Reviewed By: zzl0

Differential Revision: D48342318

fbshipit-source-id: e17ba4765316614384b8d4ebc277a15fc195cc47