formatter: give better error when write fails
Summary:
Include the underlying IO error in the error message when we fail writing command output in Rust commands.

Also print any error flushing the IO, which might be where we encounter the error due to buffering.

I tweaked the error output to use `{:#}` instead of `{}` which will the full context chain.

Reviewed By: quark-zju

Differential Revision: D48342312

fbshipit-source-id: 78e59f5bf2d04c047545626760bb8cd2a06ee43b