Implement an AWS S3 client on the S3 blob crate
Summary:
As part of the Mononoke Oprational Shadow for SEV0 scenarios, we need to make Mononoke capable of interacting with AWS. This cmmit is a first iteration of the implementation. It gets the credentials for an AWS account fron `AWSume` and hooks up to the forward proxy in order to allow writes from prod.

At some point we will modify this to work from within AWS as well, i. e. not using forward proxy, or credentials provided from Cloud Auth

This isn't yet called by anything.

Differential Revision: D48232507

fbshipit-source-id: 8a72b461224d2b0ddd22d2a99950abd01302fcba