add: accommodate behavior difference in Rust matcher
Summary:
To make things overall simpler, the Rust matchers match case insensitively on case insensitive filesystems. The Python matchers do some pre-normalization of patterns based on the dirstate, but that doesn't work in all cases and strangely couples the matcher framework to the dirstate.

Anyway, on a case insensitive filesystem in this state:

    $ sl status
    R foo
    ? FOO

If you run "sl add FOO", the Rust matcher for "FOO" matches both "foo" and "FOO", which throws off the add. Work around with some special checks for case insensitive exact match collision.

Reviewed By: zzl0

Differential Revision: D48666566

fbshipit-source-id: e00cc5c1d4d81c428b2548b9a8f2cd3b73aaf6d4