Make showing diff number configurable
Summary:
We were previously showing the diff number next to the diff badge. This is not so useful IMO.

One of our principles in ISL is to not show source control implementation details unless you need them, such as hashes and IDs, and instead prefer representing them with user-written identifies like commit titles. We were blatantly violating this for some reason here.

I had a feeling it was more useful for PR numbers than for diff numbers internally.

The only functionality provided by the diff number was to copy the number on click. You can do this from the right click menu on the diff badge. It's not such a common thing that it needs to be ALWAYS copyable.

Reviewed By: quark-zju

Differential Revision: D45712307

fbshipit-source-id: 5970099b668ea8cdf1744dd72b8df8cfab4d765c