properly update file modes during checkout
Summary:
# The problem
Users were running into issues with files not showing up as the correct type. For example, a user expected file `foo/bar` to be EXECUTABLE, but in reality it was REGULAR. This happened because of the following scenario:

Commit A has a file `foo/bar` which is a REGULAR file
Commit B updates foo/bar to EXECUTABLE

Any update from Commit A to after Commit B would cause foo/bar to exist in the working copy as a REGULAR file instead of an EXECUTABLE file.

Users would not notice this issue, the file would become materialized, and then they would accidentally commit the change that caused `foo/bar` to become a REGULAR file again (Commit C).

This commit would add even more opportunities for people to update from A -> B or B -> C or some combination, and this would add even more chances for the file's mode to become changed.

# The bug and solution
During checkout, we fail to check if a file's mode has changed. We do a check to see if the contents of the file change, but we don't check to see if the file mode has changed. This caused us to fail to update modes across commits that changed file modes.

The solution is to avoid returning early when the modes of oldScmEntry and newScmEntry differ while processing a checkout entry.

Reviewed By: xavierd

Differential Revision: D48667112

fbshipit-source-id: cb424f11cb95df46239ee49d4989baa7c869bc94