Fix commit info view getting stuck as non-editable
Summary:
This was an annoying bug: in some cases, after a while, sometimes the commit info would get stuck as non-editable, meaning you couldn't run amend or other commands and you couldn't click to edit the title/description.

The cause of this was subtle. Context: we use successions events to propagate edited commit messages forward, this ensures a commit message you've partially typed is not lost as you make other changes in your stack.

What was happening:
- perform a command like quick commit to create a new commit. Call this commit's hash A.
- click to edit that commit's message, add any content, click "amend message"
- this creates a new commit with hash A2 which is the successor of the old commit A.
- We get the results from `sl log` back which contain A2, does not contain A, and contains the predecessor info. This is detected correctly as a succession.
- Our message succession logic would then attempt to copy the existing edited commit message for A into A2. However, by the time this happens, A is no longer in the commit history, so our lookup for the pre-existing edited commit message sees no commit but that hash exists. It assumes then that this is an "optimistic" commit message (used to support creating fake commits which should not be editable).
- Now A2 has gotten the `{type: 'optimistic'}` message copied into its state. There is no way in the UI for us to clear this out, because we assume this state only happens for commits that don't exist, and we'll never load the old hash A into the commit history.

The fix is simply to never copy `type: "optimistic"` edited messages on succession.

Reviewed By: quark-zju

Differential Revision: D48748473

fbshipit-source-id: 54f48d0c17b2fafb7fcf8ca4348aa9aa2505b665