repo: resolve "main" bookmark in hoisted namespace explicitly
Summary:
When flushing the changelog, lazy changelog needs to know which parts can be
lazy. Before this change it uses something like `repo[main]`, where `main` is
usually `main` or `master`. That could cause problems with the next change,
since `main` or `master` might match commits titled `main` or `master` if the
bookmark does not exist. Update the code to limit the namespace for resolution
to avoid the issue.

Reviewed By: sggutier

Differential Revision: D48751471

fbshipit-source-id: 3d8e307425846616aa8e32e3da214bdd7499dcdb