Webview platform
Summary:
Add a new platform which supports our webview use case. This platform can be used when `sl isl` spawns a webview, and will set up the special `external.invoke` communication channel, so that the application (that spawns the webview) process can recieve messages and do actions like open external files, show dialogs, etc.

Note that this requires that the isl server is spawned with `--platform webview`. This means our application spawning code will need to know whether it intends to be used to spawn a webview or a browser, they're not interchangable (i.e. webview is not just of type "browser")

Reviewed By: quark-zju

Differential Revision: D48788661

fbshipit-source-id: 649072bf756671832aba77f29e25df4af6323cc6