Only show spinner for "resubmit stack" for the stack being submitted
Summary:
The context key for the <OperationDisabledButton> was the same for every stack, which means submitting any stack disables all stacks' buttons.

We need to use a unique key for each stack, so we know how to associate the ongoing submit command with a stack.

For "resubmit", we can simply use the diffId of the base of the stack.

For "submit entire stack", hash for the base of the stack is not sufficient. This is because the submit process may amend the commit with the diff id, which changes the hash, thus hiding the spinner halfway through the submit.

The parent hash is also not sufficient, because we may have multiple stacks rebased onto that commit.

The title is at least a passable approximation, since submitting won't change the title, and titles are often unique. We can make this slightly better by also including the parent hash.

There are _still_ situations where this is insufficient and we'll get duplicated spinners, but I consider this much better than the old situation.

Reviewed By: quark-zju

Differential Revision: D48798095

fbshipit-source-id: 3132fdaa05d058b23379b2e14a2ca9bccc19f736