ensure non-ASCII UTF-8 values are decoded correctly in Git blobs
Summary:
This fixes an issue where non-ASCII UTF-8 characters were not being decoded
correctly when reading the value of a Git blob via the GitHub API.
Because these decoded blobs are stored locally in IndexedDB, in order to for the
fix to take effect once this change goes live, you must clear out the blob cache
in IndexedDB.

This can be done in Chrome dev tools by going to **Application** and then
finding **IndexedDB** in the left-hand-nav, expanding the `github-objects-v2`
table and then clicking the **Delete database** button on the right.

Logging out and back in to reviewstack.dev will also clear the cache.

Reviewed By: vjeux

Differential Revision: D48846482

fbshipit-source-id: 039544ff837fe2948b8e30d1c349005c5f20bf3a