FileStackEditor: initial file stack editor
Summary:
Similar to the ISLv1 split editor, support moving lines between revs.
Select multiple lines to move multiple lines together.

The editing result does not save yet.

Note this does not make text editable yet, because of a conflict in line
selection behavior. If the text is editable, then the deleted lines should not
be selected (so they cannot be deleted, copied, or replaced). But we need
selection of deleted lines for movement.

Reviewed By: evangrayk

Differential Revision: D48491776

fbshipit-source-id: f476735bf0dcbd0b8d7a6a07d568ed83d9187b4f