StackEditConfirmButtons: move to a standalone module
Summary:
We're going to move the stack editing into a modal so we can show commit
(stack) and files in the same modal, to avoid confusion when you have to click
"Confirm" then "Save changes" two buttons to apply file stack editing.

Move `<StackEditConfirmButtons />` to a module so it no longer belongs to
CommitTreeList.

Reviewed By: evangrayk

Differential Revision: D48845037

fbshipit-source-id: 34a57f91acd63bdbcbc4f58b7940854ab3df52d9