Truncate list of uncommitted changes
Summary:
We had no limit to the number of uncommitted changes we show.

We should instead truncate the list after 25 files.

At the same time, let's make long lists of files scrollable, with a little fade at the bottom to hint that it's scrollable.

I experimented with having the server only return a subset of uncommitted changes, but the semantics get weird with optimistic state and selecting/deselecting files.

Instead, I think we should be ok with lots of changed files. If the number of changes is extreme (1000+), maybe we should show a top-level error. I don't do that in this diff.

We should also add a button to load more files, or even all files. Since it's scrollable now, I think this is not so bad, and we have already fetched them so it's not even async / there's no loading time.

Reviewed By: quark-zju

Differential Revision: D48812465

fbshipit-source-id: 9cf649d43d54cc4ad8bf9ee241763cb9ce35304e