pathmatcher: expand globs on Windows
Summary:
Similar to Python, the Rust pathmather machinery now expands globs within relpath patterns on Windows. This emulates typical shell expansion.

For example, `sl status *.rs` will now work on Windows in that "*.rs" will be expanded to a list of files by us (even though it isn't a glob pattern).

This only impacts the "status" command since it is the only native Rust command that uses the matcher. This behavior change is effectively behind the experimental.rustmatcher flag, which is only enabled in tests.

Reviewed By: zzl0

Differential Revision: D48807328

fbshipit-source-id: ec4f881b53687a2d16236196ed09757175a48a4f