remove workaround for submit's old metaedit
Summary:
Long ago, `jf submit` on a commit which didn't have a diff already associated with it, would submit the diff, obtain the diff number, and then metaedit the commit to add the diff number to the commit message.

This led to an odd situation: the commit hash stored in the local commits data in Phabricator did not match the commit the user had checked out.  This would trigger the `(local changes)` annotation in `hg ssl`, even though there were none.

This workaround was added to prevent this.  It would use mutation data to detect the initial metaedit, and ignore it when deciding to display `(local chnages)`.

Some time later, `jf submit` was changed to obtain the diff number first, metaedit the commit, and then submit the edited version.  This means Phabricator now has the correct commit hash.  However the workaround remained, which means the first mutation (amend/rebase/metaedit/...) after a user's initial submit does not show as `(local changes)`.

Fix this by removing the workaround.  It is no longer needed.

Reviewed By: zzl0

Differential Revision: D48681255

fbshipit-source-id: 2183f5400eb6eb37cf12a562176c1c169d9d3a28