Replace MPath with NonRootMPath
Summary:
`MPath` was created as a type in Mononoke for representing the paths of files and directories. Some of the earliest bonsai use cases of this involved `BonsaiChangesets` where the path was required to be non-root. 5 years ago we made the decision to bake-in this non-root check in the `MPath` type. Even though the type signature doesn't indicate this, `MPath` in its current form represents non-root paths, i.e. anything except just the base directory (`/`). This worked fine for cases which actually needed non-root paths but `MPath` in general have been used much more extensively throughout Mononoke codebase in the past five years. In cases where the path can also be root, we use `Option<MPath>` where the `None` variant represents the root path. In cases where such a path can be optional, we use `Option<Option<MPath>>` and it gets very ugly very quickly from there.

This diff aims to replace the `MPath` type name with `NonRootMPath` to make it more explicit what the type represents. Once landed, I will make `MPath` as a type that can represent any type of path and in particular the root would be represented as an empty vec. Finally, I will replace all occurrences of `Option<NonRootMPath>` with `MPath` wherever its safe to do so (e.g. if `Option<NonRootMPath>` is persisted in storage somewhere in a serialized form, it won't be safe to change the type)

Reviewed By: YousefSalama

Differential Revision: D48950003

fbshipit-source-id: 384e9202987dbbc608e9df378dd56d9d938c4728