Suggested rebase button
Summary:
In the old ISL, we had a button on commits that acted as a quick rebase that didn't require drag & drop, which can be cumbersome for drags over large distances.

This diff adds a conceptually similar thing, although the experience is slightly different.

Instead of a dropdown button that lets you globally choose the rebase destination, we have a generic "rebase onto..." button, which shows a context menu choice of the various public locations (remote/main, stable locations, etc).

This gives you a similarly quick way to do a rebase for older stacks. If the button doesn't appear on a specific commit, you can always drag&drop.

A couple of minor things to note:
- because we only show the suggestions on commits without any remote bookmark or stable, we don't need to filter the list shown in the context menu, as every commit is a valid dest
- because clicking a dest immediately runs the operation, which shows optimistic state that hides the original stack, there's no need to use <OperationDisabledButton /> or equivalent to show a specific contextual spinner

Reviewed By: quark-zju

Differential Revision: D48848422

fbshipit-source-id: 40376acb05712c261030f03eb9c936365b9f6f33